Centralize AI record reference formatting
