+++ packages/twenty-front/src/modules/ai/utils/formatChatReference.ts
@@ -1,8 +1,14 @@
 import { type ChatReferenceIdentity } from '@/ai/types/ChatReferenceIdentity';
 import { getChatReferenceCloseTag } from '@/ai/utils/getChatReferenceCloseTag';
 import { getChatReferenceIdentitySegment } from '@/ai/utils/getChatReferenceIdentitySegment';
+import { formatRecordReference } from 'twenty-shared/ai';
 
 export const formatChatReference = (
   reference: ChatReferenceIdentity & { displayName: string },
-): string =>
-  `[[${reference.kind}:${getChatReferenceIdentitySegment(reference)}:${reference.displayName}${getChatReferenceCloseTag(reference.kind)}`;
+): string => {
+  if (reference.kind === 'record') {
+    return formatRecordReference(reference);
+  }
+
+  return `[[${reference.kind}:${getChatReferenceIdentitySegment(reference)}:${reference.displayName}${getChatReferenceCloseTag(reference.kind)}`;
+};
