+++ packages/twenty-front/src/modules/ai/utils/formatRecordReference.ts
@@ -1,17 +1,1 @@
-import { formatChatReference } from '@/ai/utils/formatChatReference';
-
-export const formatRecordReference = ({
-  objectNameSingular,
-  recordId,
-  displayName,
-}: {
-  objectNameSingular: string;
-  recordId: string;
-  displayName: string;
-}): string =>
-  formatChatReference({
-    kind: 'record',
-    objectNameSingular,
-    recordId,
-    displayName,
-  });
+export { formatRecordReference } from 'twenty-shared/ai';
