+++ packages/twenty-shared/src/ai/utils/__tests__/format-record-reference.util.spec.ts
@@ -1,0 +1,13 @@
+import { formatRecordReference } from '@/ai/utils/format-record-reference.util';
+
+describe('formatRecordReference', () => {
+  it('formats a record reference without altering its display name', () => {
+    expect(
+      formatRecordReference({
+        objectNameSingular: 'company',
+        recordId: 'record-id',
+        displayName: '[test] ]] [test]',
+      }),
+    ).toBe('[[record:company:record-id:[test] ]] [test][[/record]]');
+  });
+});
