+++ packages/twenty-shared/src/ai/utils/format-record-reference.util.ts
@@ -1,0 +1,10 @@
+export const formatRecordReference = ({
+  objectNameSingular,
+  recordId,
+  displayName,
+}: {
+  objectNameSingular: string;
+  recordId: string;
+  displayName: string;
+}): string =>
+  `[[record:${objectNameSingular}:${recordId}:${displayName}[[/record]]`;
